Terrible block commit. Almost at working v0.
Added tests for initialization of a game, initialization of a deck.

Cards are mostly unimplemented. Going to go with a trivial deck of nothing but
VP and money for the first working version.

Players are currently raw containers for decks...need to add basic logic for v0.

Added a couple utility functions for testing equality of deck to card dict and
equality of two card dicts.